@@ -1,6 +1,6 @@
 {
   "name": "BRCAMobile",
-  "version": "1.3.14",
+  "version": "1.4.0",
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
@@ -12,38 +12,39 @@
   },
   "dependencies": {
     "babel-plugin-transform-remove-console": "^6.8.5",
-    "immutable": "^3.8.1",
+    "immutable": "^3.8.2",
     "linkify-it": "^2.0.3",
-    "lodash": "^4.17.4",
+    "lodash": "^4.17.11",
     "moment": "^2.19.2",
     "moment-timezone": "^0.5.14",
-    "prop-types": "^15.5.10",
-    "react": "16.0.0-beta.5",
-    "react-native": "0.49.0",
-    "react-native-background-fetch": "^2.0.8",
-    "react-native-background-task": "^0.2.1",
-    "react-native-fcm": "^8.0.0",
+    "prop-types": "^15.6.2",
+    "react": "16.4.1",
+    "react-native": "0.56.0",
+    "react-native-background-fetch": "^2.4.3",
     "react-native-immutable-list-view": "^0.6.2",
-    "react-native-navigation": "^1.1.287",
+    "react-native-navigation": "^1.1.486",
+    "react-native-push-notification": "^3.1.1",
     "react-native-settings-list": "^1.8.0",
-    "react-native-simple-toast": "0.0.6",
+    "react-native-simple-toast": "^0.0.8",
     "react-native-tab-view": "0.0.69",
     "react-native-vector-icons": "^4.3.0",
-    "react-redux": "^5.0.6",
+    "react-redux": "^5.0.7",
     "redux": "^3.7.2",
     "redux-batched-actions": "^0.1.6",
     "redux-immutablejs": "0.0.8",
     "redux-persist": "^5.4.0",
     "redux-persist-transform-immutable": "^4.3.0",
     "redux-thunk": "^2.2.0",
-    "whatwg-fetch": "^2.0.3"
+    "whatwg-fetch": "^2.0.4"
   },
   "devDependencies": {
     "babel-jest": "^20.0.3",
     "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react-native": "^2.1.0",
+    "babel-preset-react-native": "^5.0.0",
     "jest": "^21.2.1",
     "jest-react-native": "^18.0.0",
+    "metro": "^0.45.1",
+    "metro-core": "^0.45.1",
     "react-test-renderer": "^16.1.1"
   }
 }
