@@ -3,7 +3,8 @@ package com.brcamobile;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.jamesisaac.rnbackgroundtask.BackgroundTaskPackage;
+import com.dieam.reactnativepushnotification.ReactNativePushNotificationPackage;
+import com.transistorsoft.rnbackgroundfetch.RNBackgroundFetchPackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
@@ -24,7 +25,6 @@ import com.facebook.stetho.okhttp3.StethoInterceptor;
 import java.util.concurrent.TimeUnit;
 
 // react-native-fcm
-import com.evollu.react.fcm.FIRMessagingPackage;
 
 public class MainApplication extends NavigationApplication {
     // stetho
@@ -43,8 +43,6 @@ public class MainApplication extends NavigationApplication {
 
         // FAISAL: apparently stetho messed up this linking step
         SoLoader.init(this, /* native exopackage */ false);
-
-        BackgroundTaskPackage.useContext(this);
     }
 
     @Override
@@ -55,8 +53,8 @@ public class MainApplication extends NavigationApplication {
 
     protected List<ReactPackage> getPackages() {
         return Arrays.<ReactPackage>asList(
-            new FIRMessagingPackage(),
-            new BackgroundTaskPackage()
+            new RNBackgroundFetchPackage(),
+            new ReactNativePushNotificationPackage()
         );
     }
 
